feat: Add support for sending file responses in HttpClient
This commit adds support for sending file responses in the HttpClient class. The sendFileResponse function is implemented to open the requested file and send it to the client in chunks. If the file cannot be opened, an error message is sent instead.

The processRequest function is modified to check the HTTP method and call the appropriate method processor. If the response body is empty, indicating that the response is a file, the sendFileResponse function is called to send the file in chunks.

The sendHeader function is added to send the response headers to the client before sending the file.

The _isHeaderSent flag is introduced to track whether the response headers have been sent, and the _writingPos variable is used to keep track of the current position in the file being sent.

Note: This commit assumes that the HttpMethodProcessor functions
processGetRequest is already implemented and return the appropriate HttpResponse objects.

This commit is a work in progress and may require further refinement and testing.